requirements.txt
newsapi-python
fastapi
pydantic 
uvicorn
python-dotenv
crewai
groq